requirements.txt
streamlit
requests
python-dotenv
schedule
plotly
scikit-learn
joblib
pandas